fix(snmpmap): satisfy the repository lint sweep in the physical mapper
The media text match now looks for the shared "fib" stem so both spellings
of fiber select the arm without a word the spell linter rejects, a test
drops a redundant conversion, and the shared EtherLike fixture is used
with two different indexes so its parameter is real.

+++ src/common/snmpmap/phy_ddm_test.go
@@ -237,7 +237,7 @@ func TestPhysical_Hh3cChannelsOutOfOrder(t *testing.T) {
 	// hundredths of dBm per channel, delivered 2, 1, 4, 3
 	for _, ch := range []uint32{2, 1, 4, 3} {
 		vbs = append(vbs,
-			integerAt(hh3ctransceiverinfomib.Hh3cTransceiverChannelCurTXPower, int32(-100*int32(ch)), 49, ch),
+			integerAt(hh3ctransceiverinfomib.Hh3cTransceiverChannelCurTXPower, -100*int32(ch), 49, ch),
 			// hundredths of mA
 			integerAt(hh3ctransceiverinfomib.Hh3cTransceiverChannelBiasCurrent, 650, 49, ch),
 			// microamps
